chore(test-infra): wire socket-event logging on test HTTP server (ROK-1249, AC1)
Adds `instrumentHttpServer(server)` (new module
`api/src/common/testing/socket-debug.ts`) and invokes it from
`buildNestApp` in `test-app.ts` when `RL_TEST_SOCKET_DEBUG=true`.

Listeners cover the three failure paths the rotating-suite
`socket hang up` / ECONNRESET surfaces through:
- `request` → record method/url/startedAt on the socket via WeakMap
- `connection.error` → log code + elapsed since request start
- `connection.close hadError=true` → log elapsed since request start
- `clientError` → log err code + message

Off by default; `validate-ci.sh` and GitHub Actions are unaffected.
Also splits the snapshot helper into `snapshot-buckets.ts` so both
files stay under the 300-line cap.
